+++ src/App.tsx
@@ -1,4 +1,4 @@
-import { Grid, GridItem, Show } from '@chakra-ui/react';
+import { Grid, GridItem, HStack, Show } from '@chakra-ui/react';
 import NavBar from './components/NavBar';
 import GameGrid from './components/GameGrid';
 import GenreList from './components/GenreList';
@@ -6,6 +6,7 @@ import { useState } from 'react';
 import { Genre } from './hooks/useGenres';
 import PlatformSelector from './components/PlatformSelector';
 import { Platform } from './hooks/useGames';
+import SortSelector from './components/SortSelector';
 
 // #region Interface
 export interface GameQuery {
@@ -17,6 +18,7 @@ export interface GameQuery {
 function App() {
   // #region State
   const [gameQuery, setGameQuery] = useState<GameQuery>({} as GameQuery);
+  const [selectedSort, setSelectedSort] = useState<string>('Relevance');
   // #endregion
 
   return (
@@ -41,12 +43,17 @@ function App() {
         </GridItem>
       </Show>
       <GridItem area="main">
+        <HStack spacing={5} paddingLeft={2} marginBottom={5}>
         <PlatformSelector
           selectedPlatform={gameQuery.platform}
           onSelectPlatform={(platform) =>
             setGameQuery({ ...gameQuery, platform })
           }
         />
+        <SortSelector
+          // selectedSort={} onSelectedSort={}
+          />
+          </HStack>
         <GameGrid gameQuery={gameQuery} />
       </GridItem>
     </Grid>