+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Progress
 
 - **Finished On:** ??
+- **Add:** 26 - Building Sort Selector
 - **Add:** 25 - Extracting a Query Object
 - **Started On:** 24 - Filtering Games by Platform